refactor: 💡 line send message
