+++ src/error.ts
@@ -1,6 +1,6 @@
 import { Prisma } from "@prisma/client";
 import { AuthError } from "next-auth";
-import { sendLineNotifyMessage } from "./apis/line-notify/send-message";
+import { sendLineNotifyMessage } from "./apis/line-notify/fetch-message";
 import { ERROR_MESSAGES, FORM_ERROR_MESSAGES } from "./constants";
 import type { ServerAction } from "./types";
